Create DroneDefender
This works perfectly for my system for disabling or overtaking control of an invading drone. I believe this device can be made legal with sufficient testing. Also all the widely used Bands used by drones - non-government anyway - fall within the 2.4  and 5.8 GHz range with a few in the 900 MHz range those use something else. This utility could be used to protect not only people from peeping toms and nosy snoops, but it has potential as a way to keep Prison inmates from having items dropped by drones, government installations, as well as national defense. I have tested it with a drone and it works absolutely amazingly well. I lost full response instantly and the drone landed. The only constraint is you must keep the narrow beam pointed right at the drone constantly and that gets challenging since the beam is invisible. Wind amplifies this challenge. Once the drone is out of line of site with the operating transmitter the beam can be cut as the drone will land itself and wait for its new owners to collect it.